fixed gaps
